Alla_bots-master/requirements.txt
aiogram==2.14.1
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
Babel==2.9.0
certifi==2020.12.5
chardet==4.0.0
environs==9.3.2
idna==2.10
Markdown==3.3.4
marshmallow==3.12.1
multidict==5.1.0
python-dotenv==0.18.0
pytz==2021.1
requests==2.25.1
typing-extensions==3.7.4.3
urllib3==1.26.4
yarl==1.6.3
SQLAlchemy~=1.3.24
###psycopg2~=2.7.7
gino~=1.0.1
asgiref~=3.4.1
psycopg2-binary